refactor: Memperbaiki error pada tipe data double di vehicle table dan mengubah text pemberitahuan untuk migration file ketika sukses

+++ src/migrations/0002-create-vehicle-table.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS vehicle(
     model TEXT,
     year INT,
     mileage TEXT,
-    price DOUBLE,
+    price NUMERIC,
     condition TEXT,
     transmission TEXT,
     fuel_type TEXT,
